Big changes on 1-1-2013.

+++ Samoyan/src/samoyan/core/image/ImageSizer.java
@@ -11,5 +11,5 @@ public interface ImageSizer
 	 * @return The processed image, typically <code>src</code> after manipulation.
 	 * @throws Exception
 	 */
-	public JaiImage process(JaiImage src, int pixelRatio) throws Exception;
+	public JaiImage process(JaiImage src, float pixelRatio) throws Exception;
 }
